pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "word-counter-AI"
version = "1.0.1"
authors = [
  { name="wojpo", email="ilikeahottodogu@gmail.com" },
]
description = "AI word counter package"
dependencies = ["google-generativeai"]
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

[project.urls]
Homepage = "https://github.com/wojpo/word-counter-AI/"
Issues = "https://github.com/wojpo/word-counter-AI/issues"